basic tests working with new contract structure
